fix connection null type
